split KV publisher stages
