+++ lib/llm/src/kv_router/publisher/event_processor.rs
@@ -1,252 +1,21 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use std::collections::HashMap;
-use std::collections::hash_map::Entry;
-use std::future::Future;
 use std::sync::Arc;
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
-use anyhow::Result;
 use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
 
 use dynamo_kv_router::RouterEventSink;
 use dynamo_kv_router::indexer::LocalKvIndexer;
 use dynamo_kv_router::protocols::*;
-use dynamo_runtime::transports::event_plane::EventPublisher;
 use dynamo_runtime::transports::nats::NatsQueue;
 
-use crate::kv_router::KV_EVENT_SUBJECT;
-
+use super::batching::BatchingState;
+use super::dedup::EventDedupFilter;
+use super::sinks::{JetStreamPublisher, emit};
 use super::{DEFAULT_MAX_BATCH_BLOCKS, kv_publisher_metrics};
-
-/// Reference-counting filter that deduplicates KV cache events.
-///
-/// vLLM can emit multiple store/remove events for the same block hash.
-/// Refcounts are tracked **per DP rank** because identical block hashes
-/// on different ranks represent independent blocks.
-///
-/// - **Store**: always passes through; increments refcount for the rank.
-/// - **Remove**: only passes through when refcount decrements to 0.
-/// - **Cleared**: resets refcounts for all ranks.
-pub(super) struct EventDedupFilter {
-    /// Per-(dp_rank, storage_tier) refcounts.
-    per_rank_tier: HashMap<(u32, StorageTier), HashMap<ExternalSequenceBlockHash, usize>>,
-}
-
-impl EventDedupFilter {
-    pub(super) fn new() -> Self {
-        Self {
-            per_rank_tier: HashMap::new(),
-        }
-    }
-
-    /// Track a store event. Increments refcount for each block hash on the
-    /// given (DP rank, storage tier). Stores always pass through — this only
-    /// updates bookkeeping.
-    pub(super) fn track_store(
-        &mut self,
-        dp_rank: u32,
-        storage_tier: StorageTier,
-        data: &KvCacheStoreData,
-    ) {
-        let refcounts = self
-            .per_rank_tier
-            .entry((dp_rank, storage_tier))
-            .or_default();
-        for block in &data.blocks {
-            *refcounts.entry(block.block_hash).or_insert(0) += 1;
-        }
-    }
-
-    /// Filter a remove event. Retains only block hashes whose refcount on the
-    /// given (DP rank, storage tier) decrements to 0 (removing them from the
-    /// map). Returns `None` if no hashes survive filtering.
-    pub(super) fn filter_remove(
-        &mut self,
-        dp_rank: u32,
-        storage_tier: StorageTier,
-        mut data: KvCacheRemoveData,
-    ) -> Option<KvCacheRemoveData> {
-        let refcounts = self
-            .per_rank_tier
-            .entry((dp_rank, storage_tier))
-            .or_default();
-        data.block_hashes.retain(|hash| {
-            match refcounts.entry(*hash) {
-                Entry::Occupied(mut entry) => {
-                    *entry.get_mut() -= 1;
-                    if *entry.get() == 0 {
-                        entry.remove();
-                        true // refcount hit 0 → pass through
-                    } else {
-                        false // still has references → filter out
-                    }
-                }
-                Entry::Vacant(_) => {
-                    true // not tracked → pass through defensively
-                }
-            }
-        });
-        if data.block_hashes.is_empty() {
-            None
-        } else {
-            Some(data)
-        }
-    }
-
-    /// Clear refcounts for all DP ranks and tiers. A `Cleared` event from any
-    /// rank causes the indexer to wipe all blocks for the entire worker, so we
-    /// must reset all refcounts to stay consistent.
-    pub(super) fn clear(&mut self) {
-        self.per_rank_tier.clear();
-    }
-}
-
-/// Accumulator for in-flight KV cache events that will be merged into a single
-/// [`RouterEvent`] before being forwarded to the event sink.
-#[derive(Debug)]
-pub(super) struct BatchingState {
-    pub(super) pending_removed: Option<KvCacheRemoveData>,
-    pub(super) pending_stored: Option<KvCacheStoreData>,
-    pub(super) next_publish_id: u64,
-    pub(super) last_dp_rank: u32,
-    pub(super) last_storage_tier: StorageTier,
-    pub(super) last_flush_time: Instant,
-}
-
-impl BatchingState {
-    pub(super) fn new() -> Self {
-        Self {
-            pending_removed: None,
-            pending_stored: None,
-            next_publish_id: 1,
-            last_dp_rank: 0,
-            last_storage_tier: StorageTier::Device,
-            last_flush_time: Instant::now(),
-        }
-    }
-
-    pub(super) fn has_pending(&self) -> bool {
-        self.pending_removed.is_some() || self.pending_stored.is_some()
-    }
-
-    pub(super) fn pending_block_count(&self) -> usize {
-        self.pending_removed
-            .as_ref()
-            .map(|r| r.block_hashes.len())
-            .unwrap_or(0)
-            + self
-                .pending_stored
-                .as_ref()
-                .map(|s| s.blocks.len())
-                .unwrap_or(0)
-    }
-
-    pub(super) fn record_flush_time(&mut self) {
-        self.last_flush_time = Instant::now();
-    }
-
-    pub(super) fn remaining_timeout(&self, timeout_ms: u64) -> Duration {
-        let timeout = Duration::from_millis(timeout_ms);
-        let elapsed = self.last_flush_time.elapsed();
-        if elapsed >= timeout {
-            Duration::ZERO
-        } else {
-            timeout - elapsed
-        }
-    }
-
-    pub(super) fn is_timeout_elapsed(&self, timeout_ms: u64) -> bool {
-        self.remaining_timeout(timeout_ms) == Duration::ZERO
-    }
-
-    async fn flush<P: RouterEventSink + Send + Sync + 'static>(
-        &mut self,
-        publisher: &P,
-        local_indexer: &Option<Arc<LocalKvIndexer>>,
-        worker_id: u64,
-        dedup: &mut EventDedupFilter,
-    ) {
-        if !self.has_pending() {
-            return;
-        }
-        let dp_rank = self.last_dp_rank;
-        let mut emitted = false;
-        if let Some(data) = self.pending_removed.take()
-            && let Some(filtered) = dedup.filter_remove(dp_rank, self.last_storage_tier, data)
-        {
-            emit(
-                publisher,
-                local_indexer,
-                worker_id,
-                self.last_storage_tier,
-                KvCacheEvent {
-                    event_id: self.next_publish_id,
-                    data: KvCacheEventData::Removed(filtered),
-                    dp_rank,
-                },
-            )
-            .await;
-            emitted = true;
-        }
-        if let Some(data) = self.pending_stored.take() {
-            dedup.track_store(dp_rank, self.last_storage_tier, &data);
-            emit(
-                publisher,
-                local_indexer,
-                worker_id,
-                self.last_storage_tier,
-                KvCacheEvent {
-                    event_id: self.next_publish_id,
-                    data: KvCacheEventData::Stored(data),
-                    dp_rank,
-                },
-            )
-            .await;
-            emitted = true;
-        }
-        if emitted {
-            self.next_publish_id += 1;
-        }
-        self.record_flush_time();
-    }
-}
-
-pub(super) struct EventPlanePublisher(pub(super) EventPublisher);
-
-impl RouterEventSink for EventPlanePublisher {
-    fn publish_event(&self, event: &RouterEvent) -> impl Future<Output = Result<()>> + Send {
-        self.0.publish(event)
-    }
-}
-
-pub(super) struct JetStreamPublisher(pub(super) NatsQueue);
-
-impl RouterEventSink for JetStreamPublisher {
-    fn publish_event(&self, event: &RouterEvent) -> impl Future<Output = Result<()>> + Send {
-        NatsQueue::publish_event(&self.0, KV_EVENT_SUBJECT, event)
-    }
-}
-
-async fn emit<P: RouterEventSink>(
-    publisher: &P,
-    local_indexer: &Option<Arc<LocalKvIndexer>>,
-    worker_id: u64,
-    storage_tier: StorageTier,
-    event: KvCacheEvent,
-) {
-    let router_event = RouterEvent::with_storage_tier(worker_id, event, storage_tier);
-    if let Some(indexer) = local_indexer
-        && let Err(e) = indexer.apply_event_with_buffer(router_event.clone()).await
-    {
-        tracing::warn!(worker_id, error = %e, "Failed to apply event to local indexer");
-    }
-    if let Err(e) = publisher.publish_event(&router_event).await {
-        tracing::error!(worker_id, error = %e, "Failed to publish event");
-    }
-}
 
 pub(super) async fn run_event_processor_loop<P: RouterEventSink + Send + Sync + 'static>(
     publisher: P,
